+++ src/js/astjs.ml
@@ -186,6 +186,15 @@ module Make (Cfg : Config) = struct
       to_js_object "@" [| syntax_pos_js at.left; syntax_pos_js at.right; it |]
     else it
 
+  let add_raw_exp (exp : Syntax.exp) (it : Js.Unsafe.any) : Js.Unsafe.any =
+    (* Use Object.defineProperty to make rawExp non-enumerable, hiding it from JSON.stringify *)
+    let descriptor = Js.Unsafe.obj [|
+      ("value", Js.Unsafe.inject exp);
+      ("enumerable", Js.Unsafe.inject Js._false);
+    |] in
+    ignore (Js.Unsafe.global##.Object##defineProperty it (Js.string "rawExp") descriptor);
+    it
+
   let add_trivia (at : Source.region) (it : Js.Unsafe.any) : Js.Unsafe.any =
     match Cfg.include_docs with
     | Some table -> (
@@ -295,7 +304,7 @@ module Make (Cfg : Config) = struct
   let rec exp_js e =
     let open Syntax in
     let open Source in
-    exp'_js e |> add_type_annotation e.note.note_typ |> add_source e.at
+    exp'_js e |> add_raw_exp e |> add_type_annotation e.note.note_typ |> add_source e.at
 
   and exp'_js e =
     let open Syntax in