+++ src/js/moc_js.ml
@@ -43,5 +43,7 @@ let () =
          New versions of tooling should use [parseMotokoTypedWithScopeCache]. *)
       method parseMotokoTyped paths = js_parse_motoko_typed paths
       method parseMotokoTypedWithScopeCache enableRecovery paths scopeCache = js_parse_motoko_typed_with_scope_cache enableRecovery paths scopeCache
+      method contextualDotSuggestions scope rawExp = js_contextual_dot_suggestions scope rawExp
+      method contextualDotModule rawExp = js_contextual_dot_module rawExp
       method printDeps file = print_deps file
      end);